'main.tsのindex.htmlで使う内容をほぼeyeTrendClass.tsに移行'
