+++ client/index.html
@@ -10,8 +10,6 @@
     <link rel="stylesheet" href="src/css/top.css">
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-    <!-- jQuery CDN -->
-    <script src="https://ajax.googleapis.com/ajax/libs/jquery/2.2.4/jquery.min.js"></script>
     <title>EyeSite -TOP-</title>
   </head>
   <body>
@@ -28,6 +26,8 @@
       <button id="training" onclick="location.href=''">トレーニング</button><!-- トレーニングボタン -->
       <button id="nickname">ニックネーム</button><!-- ニックネーム欄 -->
       <canvas id="graph"></canvas><!-- グラフ表示場所 -->
+      <div id="left_result"></div>
+      <div id="right_result"></div>
     </div>
 
     <!-- 背景アニメーション -->
