+++ package.json
@@ -7,7 +7,8 @@
   "author": "takashi",
   "license": "MIT",
   "dependencies": {
-    "express": "^4.17.1"
+    "express": "^4.17.1",
+    "jquery": "^3.6.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.12",
